Moved component files into folders
