+++ src/components/JournalPage/Journal.js
