+++ src/components/JournalPage/MoodIcon.js
