+++ src/components/JournalPage/MoodTracker.js
