+++ src/components/Main/Logo.js
